+++ packages/tupaia-web/src/features/Dashboard/Dashboard.tsx
@@ -12,17 +12,16 @@ import { ErrorBoundary, SpinningLoader } from '@tupaia/ui-components';
 import { MatrixConfig } from '@tupaia/types';
 import { MOBILE_BREAKPOINT } from '../../constants';
 import { useDashboards, useEntity, useProject } from '../../api/queries';
-import { DashboardItem } from '../DashboardItem';
-import { EnlargedDashboardItem } from '../EnlargedDashboardItem';
 import { DashboardItem as DashboardItemType } from '../../types';
 import { gaEvent, getDefaultDashboard, useGAEffect } from '../../utils';
+import { DashboardItem } from '../DashboardItem';
+import { EnlargedDashboardItem } from '../EnlargedDashboardItem';
 import { ExpandButton } from './ExpandButton';
 import { Photo } from './Photo';
 import { Breadcrumbs } from './Breadcrumbs';
 import { StaticMap } from './StaticMap';
-import { DashboardMenu } from './dashboardMenu/DashboardMenu';
 import { ExportDashboard } from './ExportDashboard';
-import { SubscribeModal } from './dashboardMenu/SubscribeModal';
+import { SubscribeModal, DashboardMenu } from './DashboardMenu';
 
 const MAX_SIDEBAR_EXPANDED_WIDTH = 1000;
 const MAX_SIDEBAR_COLLAPSED_WIDTH = 550;
