+++ packages/tupaia-web/src/features/Dashboard/DashboardMenu/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+export { DashboardMenu } from './DashboardMenu';
+export { SubscribeModal } from './SubscribeModal';
