+++ packages/tupaia-web/src/features/Dashboard/ExportDashboard/ExportSubtitle.tsx
@@ -6,7 +6,7 @@
 import { Typography } from '@material-ui/core';
 import styled from 'styled-components';
 
-export const ExportSettingsInstructions = styled(Typography)`
+export const ExportSubtitle = styled(Typography)`
   color: ${({ theme }) => theme.palette.text.primary};
   font-size: 0.875rem;
   line-height: 1.4;
