+++ packages/tupaia-web/src/features/Dashboard/index.ts
@@ -4,3 +4,4 @@
  */
 
 export { Dashboard } from './Dashboard';
+export * from '../ExportSettings';
