+++ packages/tupaia-web/src/features/EnlargedDashboardItem/EnlargedDashboardItem.tsx
@@ -3,21 +3,20 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-import React, { useReducer } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { useSearchParams } from 'react-router-dom';
 import { BaseReport, ViewConfig } from '@tupaia/types';
 import { URL_SEARCH_PARAMS } from '../../constants';
 import { Modal } from '../../components';
 import { Entity } from '../../types';
+import { ExportFormats, ExportSettingsContextProvider } from '../ExportSettings';
 import { ExportDashboardItem } from './ExportDashboardItem';
 import { EnlargedDashboardVisual } from './EnlargedDashboardVisual';
 import {
-  EXPORT_FORMATS,
-  ExportContext,
-  ExportDispatchContext,
-  exportReducer,
+  ExportDashboardItemContextProvider,
   useEnlargedDashboardItem,
+  useExportDashboardItem,
 } from './utils';
 import { ExportButton } from './ExportButton';
 
@@ -38,6 +37,23 @@ const Wrapper = styled.div<{
   flex-direction: column;
 `;
 
+// This is the inner component that is rendered inside the modal, and it needs to be separate so that the context contents can be accessed
+const EnlargedDashboardItemInner = ({
+  entityName,
+  hasBigData,
+}: {
+  entityName?: Entity['name'];
+  hasBigData: boolean;
+}) => {
+  const { isExportMode } = useExportDashboardItem();
+  return (
+    <Wrapper $hasBigData={!isExportMode && hasBigData}>
+      <ExportDashboardItem entityName={entityName} />
+      <EnlargedDashboardVisual entityName={entityName} />
+    </Wrapper>
+  );
+};
+
 /**
  * EnlargedDashboardItem is the dashboard item modal. It is visible when there is a reportCode in the URL which is valid, and the dashboard item is loaded.
  */
@@ -54,16 +70,6 @@ export const EnlargedDashboardItem = ({ entityName }: { entityName?: Entity['nam
     exportWithTableDisabled = false,
   } = presentationOptions || {};
 
-  const [exportConfig, dispatch] = useReducer(exportReducer, {
-    isExporting: false,
-    isExportMode: false,
-    exportError: null,
-    exportFormat: type === 'chart' ? EXPORT_FORMATS.PNG : EXPORT_FORMATS.XLSX,
-    exportWithLabels,
-    exportWithTable,
-    exportWithTableDisabled,
-  });
-
   if (!reportCode || (!isLoadingDashboards && !currentDashboardItem)) return null;
 
   // // On close, remove the report search params from the url
@@ -74,7 +80,6 @@ export const EnlargedDashboardItem = ({ entityName }: { entityName?: Entity['nam
     setUrlSearchParams(urlSearchParams);
   };
 
-  const { isExportMode } = exportConfig;
   const isDataDownload =
     (currentDashboardItem?.config as unknown as ViewConfig)?.viewType === 'dataDownload';
 
@@ -88,15 +93,19 @@ export const EnlargedDashboardItem = ({ entityName }: { entityName?: Entity['nam
 
   return (
     <StyledModal isOpen onClose={handleCloseModal}>
-      <ExportContext.Provider value={exportConfig}>
-        <ExportDispatchContext.Provider value={dispatch}>
+      <ExportSettingsContextProvider
+        defaultSettings={{
+          exportWithLabels,
+          exportWithTable,
+          exportWithTableDisabled,
+          exportFormat: type === 'matrix' ? ExportFormats.XLSX : ExportFormats.PNG,
+        }}
+      >
+        <ExportDashboardItemContextProvider>
           <ExportButton />
-          <Wrapper $hasBigData={!isExportMode && hasBigData}>
-            <ExportDashboardItem entityName={entityName} />
-            <EnlargedDashboardVisual entityName={entityName} />
-          </Wrapper>
-        </ExportDispatchContext.Provider>
-      </ExportContext.Provider>
+          <EnlargedDashboardItemInner entityName={entityName} hasBigData={hasBigData} />
+        </ExportDashboardItemContextProvider>
+      </ExportSettingsContextProvider>
     </StyledModal>
   );
 };
