+++ packages/tupaia-web/src/features/EnlargedDashboardItem/ExportDashboardItem.tsx
@@ -3,32 +3,24 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-import React, { ChangeEvent, useContext, useRef } from 'react';
+import React, { useRef } from 'react';
 import styled from 'styled-components';
-import { FormGroup, Typography } from '@material-ui/core';
-import {
-  Button as BaseButton,
-  RadioGroup as BaseRadioGroup,
-  Checkbox as BaseCheckbox,
-  SpinningLoader,
-} from '@tupaia/ui-components';
-import {
-  ACTION_TYPES,
-  EXPORT_FORMATS,
-  ExportContext,
-  ExportDispatchContext,
-  useEnlargedDashboardItem,
-  useExportDashboardItem,
-} from './utils';
-import { Entity } from '../../types';
-import { EnlargedDashboardVisual } from './EnlargedDashboardVisual';
+import { Typography } from '@material-ui/core';
+import { Button as BaseButton, SpinningLoader } from '@tupaia/ui-components';
 import { DashboardItemTypes, ViewTypes } from '@tupaia/types';
+import { Entity } from '../../types';
+import { DisplayOptionsSettings, ExportFormatSettings, ExportFormats } from '../ExportSettings';
+import { useEnlargedDashboardItem, useExportDashboardItem } from './utils';
+import { ExportPreview } from './ExportPreview';
 
 const Wrapper = styled.div`
   flex-grow: 1;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+  fieldset {
+    margin-top: 1.5rem;
+  }
 `;
 
 const Container = styled.div`
@@ -71,46 +63,6 @@ const Button = styled(BaseButton)`
   text-transform: none;
 `;
 
-const RadioGroup = styled(BaseRadioGroup)`
-  margin-top: 1rem;
-  .MuiFormGroup-root {
-    border: none;
-    margin-top: 0.5rem;
-    flex-direction: column;
-  }
-  legend {
-    color: ${({ theme }) => theme.palette.text.primary};
-    font-size: 1rem;
-    font-weight: ${({ theme }) => theme.typography.fontWeightMedium};
-  }
-  label {
-    background: transparent;
-    border: none;
-    padding-top: 0;
-  }
-  .MuiSvgIcon-root {
-    width: 1.5rem;
-    height: 1.5rem;
-  }
-`;
-
-const Legend = styled.legend`
-  color: ${({ theme }) => theme.palette.text.primary};
-  font-size: 1rem;
-  font-weight: ${({ theme }) => theme.typography.fontWeightMedium};
-`;
-
-const Checkbox = styled(BaseCheckbox)`
-  margin: 0.5rem 0 0 1rem;
-  .MuiButtonBase-root {
-    padding: 0;
-    margin-right: 0.5rem;
-  }
-  label {
-    padding: 0.5rem 0 0 0.5rem;
-  }
-`;
-
 const RightColumn = styled.div`
   flex-grow: 1;
   width: 70%;
@@ -122,104 +74,21 @@ const RightColumn = styled.div`
   }
 `;
 
-const ScrollableContent = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  height: 25rem;
-  overflow: auto;
-  background: white;
-`;
-const PreviewWrapper = styled.div<{
-  $isPNG: boolean;
-}>`
-  height: 100%;
-  margin: 1rem 0;
-  display: ${({ $isPNG }) => ($isPNG ? 'block' : 'flex')};
-  justify-content: center;
-  align-items: center;
-  zoom: ${({ $isPNG }) => ($isPNG ? 0.5 : 1)};
-`;
-
-const PreviewContainer = styled.div`
-  min-width: 50rem; // the size of the a4 page
-  width: 100%;
-  padding: 1rem;
-  h2 {
-    color: ${({ theme }) => theme.palette.common.black};
-  }
-`;
-
-const PreviewTitle = styled(Typography).attrs({
-  variant: 'h3',
-})`
-  font-size: 1rem;
-  text-align: left;
-  font-weight: ${({ theme }) => theme.typography.fontWeightMedium};
-  margin-bottom: 1rem;
-`;
-
-const NoPreviewMessage = styled(Typography)`
-  color: ${({ theme }) => theme.palette.background.paper};
-`;
-
 export const ExportDashboardItem = ({ entityName }: { entityName?: Entity['name'] }) => {
-  const {
-    isExportMode,
-    isExporting,
-    exportFormat,
-    exportWithLabels,
-    exportWithTable,
-    exportWithTableDisabled,
-    exportError,
-  } = useContext(ExportContext);
-  const dispatch = useContext(ExportDispatchContext)!;
   const exportRef = useRef<HTMLDivElement | null>(null);
-  const { activeDashboard, currentDashboardItem, reportData } = useEnlargedDashboardItem();
-  const handleStartExport = useExportDashboardItem(
-    activeDashboard,
-    currentDashboardItem,
-    reportData,
-    entityName,
-    exportRef,
-  );
+  const { currentDashboardItem } = useEnlargedDashboardItem();
+  const { handleExport, cancelExport, isExportMode, isExporting, exportError } =
+    useExportDashboardItem(entityName, exportRef);
   if (!isExportMode) return null;
-  const cancelExport = () => {
-    dispatch({
-      type: ACTION_TYPES.SET_IS_EXPORT_MODE,
-      payload: false,
-    });
-  };
-  const setExportFormat = (e: ChangeEvent<HTMLInputElement>) => {
-    dispatch({
-      type: ACTION_TYPES.SET_EXPORT_FORMAT,
-      payload: e.target.value,
-    });
-  };
-
-  const setExportWithLabels = (e: ChangeEvent<HTMLInputElement>) => {
-    dispatch({
-      type: ACTION_TYPES.SET_EXPORT_WITH_LABELS,
-      payload: e.target.checked,
-    });
-  };
-
-  const setExportWithTable = (e: ChangeEvent<HTMLInputElement>) => {
-    dispatch({
-      type: ACTION_TYPES.SET_EXPORT_WITH_TABLE,
-      payload: e.target.checked,
-    });
-  };
 
   const exportOptions = [
     {
       label: 'PNG',
-      value: EXPORT_FORMATS.PNG,
+      value: ExportFormats.PNG,
     },
     {
       label: 'Excel (raw data)',
-      value: EXPORT_FORMATS.XLSX,
+      value: ExportFormats.XLSX,
     },
   ];
 
@@ -230,7 +99,7 @@ export const ExportDashboardItem = ({ entityName }: { entityName?: Entity['name'
   const availableExportOptions =
     isChart || viewType === ViewTypes.MultiValue
       ? exportOptions
-      : exportOptions.filter(option => option.value !== EXPORT_FORMATS.PNG);
+      : exportOptions.filter(option => option.value !== ExportFormats.PNG);
 
   return (
     <Wrapper>
@@ -242,60 +111,21 @@ export const ExportDashboardItem = ({ entityName }: { entityName?: Entity['name'
           ) : (
             <form>
               <Typography>The chart will be exported and downloaded to your browser.</Typography>
-              <RadioGroup
-                options={availableExportOptions}
-                value={exportFormat}
-                onChange={setExportFormat}
-                name="exportFormat"
-                label="Export format"
-              />
-              {exportFormat === EXPORT_FORMATS.PNG && isChart && (
-                <FormGroup>
-                  <Legend>Display options</Legend>
-                  <Checkbox
-                    label="Export with Labels"
-                    value={true}
-                    name="displayOptions"
-                    color="primary"
-                    checked={exportWithLabels}
-                    onChange={setExportWithLabels}
-                  />
-                  {!exportWithTableDisabled && (
-                    <Checkbox
-                      label="Export with Table"
-                      value={true}
-                      name="displayOptions"
-                      color="primary"
-                      checked={exportWithTable}
-                      onChange={setExportWithTable}
-                    />
-                  )}
-                </FormGroup>
-              )}
+              <ExportFormatSettings exportFormatOptions={availableExportOptions} />
+              {isChart && <DisplayOptionsSettings />}
             </form>
           )}
           {exportError && <Typography color="error">{exportError}</Typography>}
         </LeftColumn>
         <RightColumn>
-          <PreviewTitle>Preview</PreviewTitle>
-          <ScrollableContent>
-            <PreviewWrapper $isPNG={exportFormat === EXPORT_FORMATS.PNG}>
-              {exportFormat === EXPORT_FORMATS.PNG ? (
-                <PreviewContainer ref={exportRef}>
-                  <EnlargedDashboardVisual entityName={entityName} isPreview />
-                </PreviewContainer>
-              ) : (
-                <NoPreviewMessage>No preview available</NoPreviewMessage>
-              )}
-            </PreviewWrapper>
-          </ScrollableContent>
+          <ExportPreview entityName={entityName} exportRef={exportRef} />
         </RightColumn>
       </ExportContentContainer>
       <ButtonWrapper>
         <Button onClick={cancelExport} variant="outlined" color="default" disabled={isExporting}>
           Cancel
         </Button>
-        <Button color="primary" onClick={handleStartExport} disabled={isExporting}>
+        <Button color="primary" onClick={handleExport} disabled={isExporting}>
           Export chart
         </Button>
       </ButtonWrapper>
