+++ packages/tupaia-web/src/features/EnlargedDashboardItem/index.ts
@@ -3,3 +3,4 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 export { EnlargedDashboardItem } from './EnlargedDashboardItem';
+export { EnlargedDashboardVisual } from './EnlargedDashboardVisual';
