+++ packages/tupaia-web/src/features/ExportSettings/ExportFormatSettings.tsx
@@ -1,0 +1,41 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+
+import React from 'react';
+import styled from 'styled-components';
+import {
+  FormControl,
+  RadioGroup,
+  FormControlLabel as BaseFormControlLabel,
+  Radio,
+} from '@material-ui/core';
+import { ExportSettingLabel, useExportSettings } from '.';
+
+const FormControlLabel = styled(BaseFormControlLabel)`
+  margin-left: 0;
+  margin-top: 0.625rem;
+  .MuiButtonBase-root {
+    padding: 0 0.8rem 0 0;
+  }
+`;
+
+export const ExportFormatSettings = ({ exportFormatOptions }) => {
+  const { exportFormat, setExportFormat } = useExportSettings();
+  return (
+    <FormControl component="fieldset">
+      <ExportSettingLabel as="legend">Export format</ExportSettingLabel>
+      <RadioGroup name="exportFormat" value={exportFormat} onChange={setExportFormat}>
+        {exportFormatOptions.map(exportFormatOption => (
+          <FormControlLabel
+            value={exportFormatOption.value}
+            control={<Radio color="primary" />}
+            label={exportFormatOption.label}
+            key={exportFormatOption.value}
+          />
+        ))}
+      </RadioGroup>
+    </FormControl>
+  );
+};
