+++ packages/tupaia-web/src/features/ExportSettings/ExportSettingLabel.tsx
@@ -3,9 +3,10 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
+import { Typography } from '@material-ui/core';
 import styled from 'styled-components';
 
-export const ExportSettingLabel = styled.legend`
+export const ExportSettingLabel = styled(Typography)`
   color: ${({ theme }) => theme.palette.text.primary};
   padding-inline-start: 0;
   font-size: 1.125rem;
